test: update test imports for consolidated route paths
Fix test imports pointing to deleted route subdirectories. Update health
pipeline test imports from @/app/api/health/pipeline/route to
@/app/api/health/route. Update cron, user, and other test imports.
