+++ __tests__/api/health/pipeline-health-connection-pool.test.ts
@@ -127,7 +127,7 @@ jest.mock('@/lib/db/supabase-config', () => ({
   }),
 }));
 
-import { GET, clearHealthCache, resetOrphanSampleCounter } from '@/app/api/health/pipeline/route';
+import { GET, clearHealthCache, resetOrphanSampleCounter } from '@/app/api/health/route';
 
 function createMockRequest(): NextRequest {
   return new NextRequest('http://localhost:3000/api/health/pipeline?bypass-cache=true', {
