+++ __tests__/api/health/pipeline-health-orphan-sampling.test.ts
@@ -81,7 +81,7 @@ jest.mock('@/lib/db/supabase-config', () => ({
   }),
 }));
 
-import { GET, clearHealthCache, resetOrphanSampleCounter } from '@/app/api/health/pipeline/route';
+import { GET, clearHealthCache, resetOrphanSampleCounter } from '@/app/api/health/route';
 
 function createMockRequest(): NextRequest {
   return new NextRequest('http://localhost:3000/api/health/pipeline?bypass-cache=true', {
