+++ tests/security/auth-bypass-prevention.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { NextRequest } from 'next/server';
-import { GET } from '../../app/api/cron/tier-aware/route';
+import { GET } from '../../app/api/cron/route';
 import { HmacAuthService } from '../../lib/security/hmac-auth';
 
 // Mock dependencies similar to cron-security.test.ts
